refactor:#11 cleaning legacy, path
