--- legacy/model/config/kobert.json
@@ -1,32 +1,0 @@
-{
-    "data_dir": "data/processed/",
-    "category_dir": "data/processed/category_set.json",
-    "ckpt_dir": "model/ckpt/",
-    "train_file": "train.jsonl",
-    "dev_file": null,
-    "test_file": "test.jsonl",
-    "evaluate_test_during_training": true,
-    "eval_all_ckpt": false,
-    "save_optimizer": false,
-    "do_lower_case": false,
-    "max_seq_len": 128,
-    "num_train_epochs": 5,
-    "weight_decay": 0.0,
-    "gradient_accumulation_steps": 1,
-    "adam_epsilon": 1e-8,
-    "warmup_proportion": 0,
-    "max_steps": -1,
-    "max_grad_norm": 1.0,
-    "no_cuda": false,
-    "model_type": "kobert",
-    "model_name_or_path": "monologg/kobert",
-    "output_dir": "model/processed/kobert/",
-    "seed": 42,
-    "train_batch_size": 32,
-    "eval_batch_size": 128,
-    "logging_steps": 1000,
-    "save_steps": 1000,
-    "learning_rate": 5e-5,
-    "kfold_num": 5,
-    "sentiments" : ["pos", "neg", "none"]
-}
