--- legacy/model/src/__init__.py
@@ -1,5 +1,0 @@
-from .train import Trainer
-from .evaluation import Evaluator
-from .test import Tester
-
-__all__ = ['Trainer', 'Evaluator', 'Tester']
